requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1

environs~=9.3.2
django-filter~=2.4.0